minor: changed powermock tests to normal tests
