+++ src/test/java/com/puppycrawl/tools/checkstyle/PackageNamesLoaderTest.java
@@ -20,15 +20,32 @@
 package com.puppycrawl.tools.checkstyle;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotEquals;
+import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
+import java.io.File;
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+import java.net.URLConnection;
+import java.net.URLStreamHandler;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Enumeration;
+import java.util.HashSet;
 import java.util.Set;
 
 import org.junit.Test;
+import org.xml.sax.SAXException;
 
 import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
 
 /**
- * Enter a description of class PackageNamesLoaderTest.java.
+ * Custom class loader is needed to pass URLs to pretend these are loaded from the classpath
+ * though we can't add/change the files for testing. The class loader is nested in this class,
+ * so the custom class loader we are using is safe.
+ * @noinspection ClassLoaderInstantiation
  */
 public class PackageNamesLoaderTest extends AbstractPathTestSupport {
 
@@ -45,6 +62,176 @@ public class PackageNamesLoaderTest extends AbstractPathTestSupport {
                         .getContextClassLoader());
         assertEquals("pkgNames.length.", 0,
                 packageNames.size());
+    }
+
+    @Test
+    public void testNoPackages() throws Exception {
+        final Set<String> actualPackageNames = PackageNamesLoader
+                .getPackageNames(new TestUrlsClassLoader(Collections.emptyEnumeration()));
+
+        assertEquals("Invalid package names length.", 0, actualPackageNames.size());
+    }
+
+    @Test
+    public void testPackagesFile() throws Exception {
+        final Enumeration<URL> enumeration = Collections.enumeration(Collections.singleton(
+                new File(getPath("InputPackageNamesLoaderFile.xml")).toURI().toURL()));
+
+        final Set<String> actualPackageNames = PackageNamesLoader
+                .getPackageNames(new TestUrlsClassLoader(enumeration));
+        final String[] expectedPackageNames = {
+            "com.puppycrawl.tools.checkstyle",
+            "com.puppycrawl.tools.checkstyle.checks",
+            "com.puppycrawl.tools.checkstyle.checks.annotation",
+            "com.puppycrawl.tools.checkstyle.checks.blocks",
+            "com.puppycrawl.tools.checkstyle.checks.coding",
+            "com.puppycrawl.tools.checkstyle.checks.design",
+            "com.puppycrawl.tools.checkstyle.checks.header",
+            "com.puppycrawl.tools.checkstyle.checks.imports",
+            "com.puppycrawl.tools.checkstyle.checks.indentation",
+            "com.puppycrawl.tools.checkstyle.checks.javadoc",
+            "com.puppycrawl.tools.checkstyle.checks.metrics",
+            "com.puppycrawl.tools.checkstyle.checks.modifier",
+            "com.puppycrawl.tools.checkstyle.checks.naming",
+            "com.puppycrawl.tools.checkstyle.checks.regexp",
+            "com.puppycrawl.tools.checkstyle.checks.sizes",
+            "com.puppycrawl.tools.checkstyle.checks.whitespace",
+            "com.puppycrawl.tools.checkstyle.filefilters",
+            "com.puppycrawl.tools.checkstyle.filters",
+        };
+
+        assertEquals("Invalid package names length.", expectedPackageNames.length,
+            actualPackageNames.size());
+        final Set<String> checkstylePackagesSet =
+                new HashSet<>(Arrays.asList(expectedPackageNames));
+        assertEquals("Invalid names set.", checkstylePackagesSet, actualPackageNames);
+    }
+
+    @Test
+    public void testPackagesWithDots() throws Exception {
+        final Enumeration<URL> enumeration = Collections.enumeration(Collections.singleton(
+                new File(getPath("InputPackageNamesLoaderWithDots.xml")).toURI().toURL()));
+
+        final Set<String> actualPackageNames = PackageNamesLoader
+                .getPackageNames(new TestUrlsClassLoader(enumeration));
+        final String[] expectedPackageNames = {
+            "coding.",
+        };
+
+        assertEquals("Invalid package names length.", expectedPackageNames.length,
+            actualPackageNames.size());
+        final Set<String> checkstylePackagesSet =
+                new HashSet<>(Arrays.asList(expectedPackageNames));
+        assertEquals("Invalid names set.", checkstylePackagesSet, actualPackageNames);
+    }
+
+    @Test
+    public void testPackagesWithDotsEx() throws Exception {
+        final Enumeration<URL> enumeration = Collections.enumeration(Collections.singleton(
+                new File(getPath("InputPackageNamesLoaderWithDotsEx.xml")).toURI().toURL()));
+
+        final Set<String> actualPackageNames = PackageNamesLoader
+                .getPackageNames(new TestUrlsClassLoader(enumeration));
+        final String[] expectedPackageNames = {
+            "coding.specific",
+            "coding.",
+        };
+
+        assertEquals("Invalid package names length.", expectedPackageNames.length,
+            actualPackageNames.size());
+        final Set<String> checkstylePackagesSet =
+                new HashSet<>(Arrays.asList(expectedPackageNames));
+        assertEquals("Invalid names set.", checkstylePackagesSet, actualPackageNames);
+    }
+
+    @Test
+    public void testPackagesWithSaxException() throws Exception {
+        final Enumeration<URL> enumeration = Collections.enumeration(Collections.singleton(
+                new File(getPath("InputPackageNamesLoaderNotXml.java")).toURI().toURL()));
+
+        try {
+            PackageNamesLoader.getPackageNames(new TestUrlsClassLoader(enumeration));
+            fail("CheckstyleException is expected");
+        }
+        catch (CheckstyleException ex) {
+            assertTrue("Invalid exception cause class", ex.getCause() instanceof SAXException);
+        }
+    }
+
+    @Test
+    public void testPackagesWithIoException() throws Exception {
+        final URLConnection urlConnection = new URLConnection(null) {
+            @Override
+            public void connect() {
+                // no code
+            }
+
+            @Override
+            public InputStream getInputStream() {
+                return null;
+            }
+        };
+        final URL url = new URL("test", null, 0, "", new URLStreamHandler() {
+            @Override
+            protected URLConnection openConnection(URL u) {
+                return urlConnection;
+            }
+        });
+
+        final Enumeration<URL> enumeration = Collections.enumeration(Collections.singleton(url));
+
+        try {
+            PackageNamesLoader.getPackageNames(new TestUrlsClassLoader(enumeration));
+            fail("CheckstyleException is expected");
+        }
+        catch (CheckstyleException ex) {
+            assertTrue("Invalid exception cause class", ex.getCause() instanceof IOException);
+            assertNotEquals("Invalid exception message",
+                    "unable to get package file resources", ex.getMessage());
+        }
+    }
+
+    @Test
+    public void testPackagesWithIoExceptionGetResources() {
+        try {
+            PackageNamesLoader.getPackageNames(new TestIoExceptionClassLoader());
+            fail("CheckstyleException is expected");
+        }
+        catch (CheckstyleException ex) {
+            assertTrue("Invalid exception cause class", ex.getCause() instanceof IOException);
+            assertEquals("Invalid exception message",
+                    "unable to get package file resources", ex.getMessage());
+        }
+    }
+
+    /**
+     * Custom class loader is needed to pass URLs to pretend these are loaded from the classpath
+     * though we can't add/change the files for testing.
+     * @noinspection CustomClassloader
+     */
+    private static class TestUrlsClassLoader extends ClassLoader {
+
+        private final Enumeration<URL> urls;
+
+        TestUrlsClassLoader(Enumeration<URL> urls) {
+            this.urls = urls;
+        }
+
+        @Override
+        public Enumeration<URL> getResources(String name) {
+            return urls;
+        }
+    }
+
+    /**
+     * Custom class loader is needed to throw an exception to test a catch statement.
+     * @noinspection CustomClassloader
+     */
+    private static class TestIoExceptionClassLoader extends ClassLoader {
+        @Override
+        public Enumeration<URL> getResources(String name) throws IOException {
+            throw new IOException("test");
+        }
     }
 
 }
