+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/NewlineAtEndOfFileCheckTest.java
@@ -21,17 +21,21 @@ package com.puppycrawl.tools.checkstyle.checks;
 
 import static com.puppycrawl.tools.checkstyle.checks.NewlineAtEndOfFileCheck.MSG_KEY_NO_NEWLINE_EOF;
 import static com.puppycrawl.tools.checkstyle.checks.NewlineAtEndOfFileCheck.MSG_KEY_UNABLE_OPEN;
+import static java.util.Locale.ENGLISH;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.io.File;
+import java.io.IOException;
+import java.io.RandomAccessFile;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
 import org.junit.Test;
+import org.powermock.reflect.Whitebox;
 
 import com.puppycrawl.tools.checkstyle.AbstractModuleTestSupport;
 import com.puppycrawl.tools.checkstyle.DefaultConfiguration;
@@ -192,6 +196,32 @@ public class NewlineAtEndOfFileCheckTest
         final Iterator<LocalizedMessage> iterator = messages.iterator();
         assertEquals("Violation message differs from expected",
                 getCheckMessage(MSG_KEY_UNABLE_OPEN, ""), iterator.next().getMessage());
+    }
+
+    @Test
+    public void testWrongSeparatorLength() throws Exception {
+        final RandomAccessFile file = new RandomAccessFile(
+                getPath("InputNewlineAtEndOfFileLf.java"), "r") {
+            @Override
+            public int read(byte[] bytes) {
+                return 0;
+            }
+        };
+
+        try {
+            Whitebox.invokeMethod(new NewlineAtEndOfFileCheck(), "endsWithNewline", file);
+            fail("Exception is expected");
+        }
+        catch (IOException ex) {
+            if (System.getProperty("os.name").toLowerCase(ENGLISH).startsWith("windows")) {
+                assertEquals("Error message is unexpected",
+                        "Unable to read 2 bytes, got 0", ex.getMessage());
+            }
+            else {
+                assertEquals("Error message is unexpected",
+                        "Unable to read 1 bytes, got 0", ex.getMessage());
+            }
+        }
     }
 
 }
