+++ src/test/java/com/puppycrawl/tools/checkstyle/checks/SuppressWarningsHolderTest.java
@@ -21,24 +21,33 @@ package com.puppycrawl.tools.checkstyle.checks;
 
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.io.File;
+import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 
+import org.junit.After;
 import org.junit.Test;
+import org.powermock.reflect.Whitebox;
 
 import com.puppycrawl.tools.checkstyle.AbstractModuleTestSupport;
+import com.puppycrawl.tools.checkstyle.Checker;
 import com.puppycrawl.tools.checkstyle.DefaultConfiguration;
 import com.puppycrawl.tools.checkstyle.JavaParser;
 import com.puppycrawl.tools.checkstyle.TreeWalker;
+import com.puppycrawl.tools.checkstyle.api.AuditEvent;
 import com.puppycrawl.tools.checkstyle.api.Configuration;
 import com.puppycrawl.tools.checkstyle.api.DetailAST;
+import com.puppycrawl.tools.checkstyle.api.LocalizedMessage;
 import com.puppycrawl.tools.checkstyle.api.TokenTypes;
+import com.puppycrawl.tools.checkstyle.checks.naming.MemberNameCheck;
 import com.puppycrawl.tools.checkstyle.checks.whitespace.AbstractParenPadCheck;
 import com.puppycrawl.tools.checkstyle.checks.whitespace.TypecastParenPadCheck;
 import com.puppycrawl.tools.checkstyle.filters.SuppressWarningsFilter;
@@ -50,6 +59,17 @@ public class SuppressWarningsHolderTest extends AbstractModuleTestSupport {
     @Override
     protected String getPackageLocation() {
         return "com/puppycrawl/tools/checkstyle/checks/suppresswarningsholder";
+    }
+
+    @After
+    public void cleanUp() {
+        // clear cache that may have been set by tests
+
+        new SuppressWarningsHolder().beginTree(null);
+
+        final Map<String, String> map = Whitebox.getInternalState(SuppressWarningsHolder.class,
+                "CHECK_ALIAS_MAP");
+        map.clear();
     }
 
     @Test
@@ -149,6 +169,83 @@ public class SuppressWarningsHolderTest extends AbstractModuleTestSupport {
             assertEquals("Error message is unexpected",
                     "'=' expected in alias list item: =SomeAlias", ex.getMessage());
         }
+    }
+
+    @Test
+    public void testIsSuppressed() throws Exception {
+        populateHolder("MockEntry", 100, 100, 350, 350);
+        final AuditEvent event = createAuditEvent("check", 100, 10);
+
+        assertFalse("Event is not suppressed", SuppressWarningsHolder.isSuppressed(event));
+    }
+
+    @Test
+    public void testIsSuppressedByName() throws Exception {
+        populateHolder("check", 100, 100, 350, 350);
+        final SuppressWarningsHolder holder = new SuppressWarningsHolder();
+        final AuditEvent event = createAuditEvent("id", 110, 10);
+        holder.setAliasList(MemberNameCheck.class.getName() + "=check");
+
+        assertTrue("Event is not suppressed", SuppressWarningsHolder.isSuppressed(event));
+    }
+
+    @Test
+    public void testIsSuppressedByModuleId() throws Exception {
+        populateHolder("check", 100, 100, 350, 350);
+        final AuditEvent event = createAuditEvent("check", 350, 350);
+
+        assertTrue("Event is not suppressed", SuppressWarningsHolder.isSuppressed(event));
+    }
+
+    @Test
+    public void testIsSuppressedAfterEventEnd() throws Exception {
+        populateHolder("check", 100, 100, 350, 350);
+        final AuditEvent event = createAuditEvent("check", 350, 352);
+
+        assertFalse("Event is not suppressed", SuppressWarningsHolder.isSuppressed(event));
+    }
+
+    @Test
+    public void testIsSuppressedAfterEventEnd2() throws Exception {
+        populateHolder("check", 100, 100, 350, 350);
+        final AuditEvent event = createAuditEvent("check", 400, 10);
+
+        assertFalse("Event is not suppressed", SuppressWarningsHolder.isSuppressed(event));
+    }
+
+    @Test
+    public void testIsSuppressedAfterEventStart() throws Exception {
+        populateHolder("check", 100, 100, 350, 350);
+        final AuditEvent event = createAuditEvent("check", 100, 100);
+
+        assertTrue("Event is not suppressed", SuppressWarningsHolder.isSuppressed(event));
+    }
+
+    @Test
+    public void testIsSuppressedWithAllArgument() throws Exception {
+        populateHolder("all", 100, 100, 350, 350);
+
+        final Checker source = new Checker();
+        final LocalizedMessage firstMessageForTest =
+            new LocalizedMessage(100, 10, null, null, null, "id", MemberNameCheck.class, "msg");
+        final AuditEvent firstEventForTest =
+            new AuditEvent(source, "fileName", firstMessageForTest);
+        assertFalse("Event is suppressed",
+                SuppressWarningsHolder.isSuppressed(firstEventForTest));
+
+        final LocalizedMessage secondMessageForTest =
+            new LocalizedMessage(100, 150, null, null, null, "id", MemberNameCheck.class, "msg");
+        final AuditEvent secondEventForTest =
+            new AuditEvent(source, "fileName", secondMessageForTest);
+        assertTrue("Event is not suppressed",
+                SuppressWarningsHolder.isSuppressed(secondEventForTest));
+
+        final LocalizedMessage thirdMessageForTest =
+            new LocalizedMessage(200, 1, null, null, null, "id", MemberNameCheck.class, "msg");
+        final AuditEvent thirdEventForTest =
+            new AuditEvent(source, "fileName", thirdMessageForTest);
+        assertTrue("Event is not suppressed",
+                SuppressWarningsHolder.isSuppressed(thirdEventForTest));
     }
 
     @Test
@@ -300,6 +397,30 @@ public class SuppressWarningsHolderTest extends AbstractModuleTestSupport {
             TestUtil.isStatefulFieldClearedDuringBeginTree(check, annotationDef.get(),
                 "ENTRIES",
                 entries -> ((ThreadLocal<List<Object>>) entries).get().isEmpty()));
+    }
+
+    private static void populateHolder(String checkName, int firstLine,
+                                                         int firstColumn, int lastLine,
+                                                         int lastColumn) throws Exception {
+        final Class<?> entry = Class
+                .forName("com.puppycrawl.tools.checkstyle.checks.SuppressWarningsHolder$Entry");
+        final Constructor<?> entryConstr = entry.getDeclaredConstructor(String.class, int.class,
+                int.class, int.class, int.class);
+        entryConstr.setAccessible(true);
+
+        final Object entryInstance = entryConstr.newInstance(checkName, firstLine,
+                firstColumn, lastLine, lastColumn);
+
+        final ThreadLocal<List<Object>> entries = Whitebox
+                .getInternalState(SuppressWarningsHolder.class, "ENTRIES");
+        entries.get().add(entryInstance);
+    }
+
+    private static AuditEvent createAuditEvent(String moduleId, int line, int column) {
+        final Checker source = new Checker();
+        final LocalizedMessage message = new LocalizedMessage(line, column, null, null, null,
+                moduleId, MemberNameCheck.class, "message");
+        return new AuditEvent(source, "filename", message);
     }
 
 }
