+++ src/test/java/com/puppycrawl/tools/checkstyle/internal/powermock/HeaderCheckPowerTest.java
@@ -25,8 +25,6 @@ import static org.junit.Assert.fail;
 import static org.mockito.ArgumentMatchers.any;
 
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -35,11 +33,9 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 import com.puppycrawl.tools.checkstyle.AbstractModuleTestSupport;
-import com.puppycrawl.tools.checkstyle.api.CheckstyleException;
 import com.puppycrawl.tools.checkstyle.checks.header.AbstractHeaderCheck;
 import com.puppycrawl.tools.checkstyle.checks.header.HeaderCheck;
 import com.puppycrawl.tools.checkstyle.checks.header.HeaderCheckTest;
-import com.puppycrawl.tools.checkstyle.utils.CommonUtil;
 
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({ HeaderCheck.class, HeaderCheckTest.class, AbstractHeaderCheck.class })
@@ -50,6 +46,11 @@ public class HeaderCheckPowerTest extends AbstractModuleTestSupport {
         return "com/puppycrawl/tools/checkstyle/checks/header/header";
     }
 
+    /**
+     * This test needs powermock because {@code StringReader} can't throw an exception to trigger
+     * the catch otherwise unless the reader is mishandled.
+     * @throws Exception if there is an error.
+     */
     @Test
     public void testIoExceptionWhenLoadingHeader() throws Exception {
         final HeaderCheck check = PowerMockito.spy(new HeaderCheck());
@@ -63,27 +64,6 @@ public class HeaderCheckPowerTest extends AbstractModuleTestSupport {
         catch (IllegalArgumentException ex) {
             assertTrue("Invalid exception cause", ex.getCause() instanceof IOException);
             assertEquals("Invalid exception message", "unable to load header", ex.getMessage());
-        }
-    }
-
-    @Test
-    public void testIoExceptionWhenLoadingHeaderFile() throws Exception {
-        final HeaderCheck check = PowerMockito.spy(new HeaderCheck());
-        PowerMockito.doThrow(new IOException("expected exception")).when(check, "loadHeader",
-                any());
-
-        check.setHeaderFile(CommonUtil.getUriByFilename(getPath("InputHeaderRegexp.java")));
-
-        final Method loadHeaderFile = AbstractHeaderCheck.class.getDeclaredMethod("loadHeaderFile");
-        loadHeaderFile.setAccessible(true);
-        try {
-            loadHeaderFile.invoke(check);
-            fail("Exception expected");
-        }
-        catch (InvocationTargetException ex) {
-            assertTrue("Invalid exception cause", ex.getCause() instanceof CheckstyleException);
-            assertTrue("Invalid exception cause message",
-                ex.getCause().getMessage().startsWith("unable to load header file "));
         }
     }
