+++ src/test/resources/com/puppycrawl/tools/checkstyle/packagenamesloader/InputPackageNamesLoaderNotXml.java
@@ -1,0 +1,4 @@
+package com.puppycrawl.tools.checkstyle.packagenamesloader;
+
+public class InputPackageNamesLoaderNotXml {
+}
