+++ src/test/resources/com/puppycrawl/tools/checkstyle/packagenamesloader/InputPackageNamesLoaderWithDotsEx.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!DOCTYPE checkstyle-packages PUBLIC
+    "-//Checkstyle//DTD Package Names Configuration 1.0//EN"
+    "https://checkstyle.org/dtds/packages_1_0.dtd">
+
+<checkstyle-packages>
+  <package name="coding.">
+    <package name="specific"/>
+  </package>
+</checkstyle-packages>
